refactor chat listeners
