+++ app/src/main/java/naoned/sil/lp/naonedchat/listeners/chat/MessageListener.java
@@ -5,6 +5,6 @@ import org.jivesoftware.smack.packet.Message;
 /**
  * Created by ACHP on 24/01/2016.
  */
-public interface onMessageListener {
+public interface MessageListener {
     public void onNewMessage(Message message);
 }
